♻️ Refactor: 이미지 저장과 공유 카드 통합
